# docs/openapi/public-openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "x-ai-gateway Public API",
    "version" : "2026.05.21",
    "description" : "公开接入面的最小 OpenAPI 事实源，按头部 native provider 和 Lossless Translation Matrix 覆盖 docs、OpenAI-compatible、Claude/Gemini/Vertex 功能性入口、Codex Responses smoke 边界和 Media provider matrix；不可无损翻译的资源属性直接失败。"
  },
  "servers" : [ {
    "url" : "https://gateway.example.com",
    "description" : "生产网关地址示例"
  } ],
  "paths" : {
...
    "/v1/responses" : {
      "post" : {
        "summary" : "OpenAI-compatible Responses",
        "operationId" : "post_v1_responses",
        "tags" : [ "gateway" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "model" ],
                "properties" : {
                  "model" : {
                    "type" : "string",
                    "description" : "OpenAI model name or gateway model alias."
                  },
                  "input" : {
                    "type" : "object",
                    "description" : "Responses input string, message array or item array."
                  },
                  "stream" : {
                    "type" : "boolean",
                    "description" : "When true, emit OpenAI Responses-style server-sent events. OpenAI Direct streams pass through upstream raw SSE events when native passthrough is available."
                  },
                  "stream_options" : {
                    "type" : "object",
                    "description" : "Responses streaming options such as include_obfuscation; obfuscation is included on delta events by default and can be disabled with include_obfuscation=false."
                  },
                  "store" : {
                    "type" : "boolean",
                    "description" : "When true, persist the Response as a local resp_ resource."
                  },
                  "background" : {
                    "type" : "boolean",
                    "description" : "When true, mark stored local Response as background-capable for cancel."
                  },
                  "metadata" : {
                    "type" : "object",
                    "description" : "OpenAI metadata object preserved by the gateway."
                  },
                  "include" : {
                    "type" : "array",
                    "description" : "Additional output fields requested by OpenAI Responses clients."
                  },
                  "previous_response_id" : {
                    "type" : "string",
                    "description" : "Previous Response id for conversation continuation."
                  },
                  "tools" : {
                    "type" : "array",
                    "description" : "Responses tool definitions. Function tools execute through canonical tool calling; hosted file_search requires OpenAI Direct native execution and otherwise fails with native_hosted_tool_required; other hosted/MCP/custom tools are rejected explicitly instead of being ignored."
                  },
                  "tool_choice" : {
                    "type" : "object",
                    "description" : "Responses tool choice string or object. Non-function forced tool choices are rejected explicitly until their execution boundary is implemented."
                  }
                }
              }
            }
          }
        },
        "parameters" : [ {
          "name" : "Idempotency-Key",
          "in" : "header",
          "required" : false,
          "description" : "Optional key for non-streaming response replay.",
          "schema" : {
            "type" : "string"
          }
        } ]
      }
    },
...
    "/v1/audio/transcriptions" : {
      "post" : {
        "summary" : "Create audio transcription for multimodal input",
        "operationId" : "post_v1_audio_transcriptions",
        "tags" : [ "gateway" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "description" : "Transcribes an uploaded audio file through provider-governed routing as part of the multimodal conversation support surface.",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "required" : [ "file", "model" ],
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary",
                    "description" : "Audio file to transcribe."
                  },
                  "model" : {
                    "type" : "string",
                    "description" : "Transcription model name or gateway model alias."
                  },
                  "language" : {
                    "type" : "string",
                    "description" : "Optional input language hint."
                  },
                  "prompt" : {
                    "type" : "string",
                    "description" : "Optional transcription prompt."
                  },
                  "response_format" : {
                    "type" : "string",
                    "description" : "Optional response format when supported by the routed provider."
                  },
                  "temperature" : {
                    "type" : "string",
                    "description" : "Optional temperature, accepted as multipart text."
                  },
                  "timestamp_granularities[]" : {
                    "type" : "array",
                    "description" : "Optional timestamp granularity hints preserved as multipart text values."
                  },
                  "include[]" : {
                    "type" : "array",
                    "description" : "Optional extra response fields preserved as multipart text values when supported by the routed provider."
                  },
                  "stream" : {
                    "type" : "string",
                    "description" : "Optional streaming flag accepted as multipart text; provider support is route-dependent."
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/audio/speech" : {
      "post" : {
        "summary" : "Create speech audio for multimodal output",
        "operationId" : "post_v1_audio_speech",
        "tags" : [ "gateway" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "description" : "Generates speech audio through provider-governed routing for multimodal response output.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "model", "input", "voice" ],
                "properties" : {
                  "model" : {
                    "type" : "string",
                    "description" : "Speech model name or gateway model alias."
                  },
                  "input" : {
                    "type" : "string",
                    "description" : "Text input used to generate speech audio."
                  },
                  "voice" : {
                    "type" : "string",
                    "description" : "Voice identifier supported by the routed provider."
                  },
                  "response_format" : {
                    "type" : "string",
                    "description" : "Optional audio format such as mp3, wav or opus."
                  },
                  "speed" : {
                    "type" : "number",
                    "description" : "Optional playback speed hint when supported."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/images/generations" : {
      "post" : {
        "summary" : "Create image generation for multimodal output",
        "operationId" : "post_v1_images_generations",
        "tags" : [ "gateway" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "description" : "Creates image generations through provider-governed routing for the multimodal output surface.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "prompt" ],
                "properties" : {
                  "model" : {
                    "type" : "string",
                    "description" : "Image model name or gateway model alias."
                  },
                  "prompt" : {
                    "type" : "string",
                    "description" : "Image generation prompt."
                  },
                  "size" : {
                    "type" : "string",
                    "description" : "Optional image size supported by the routed provider."
                  },
                  "quality" : {
                    "type" : "string",
                    "description" : "Optional quality hint supported by the routed provider."
                  },
                  "n" : {
                    "type" : "integer",
                    "description" : "Optional number of images."
                  },
                  "response_format" : {
                    "type" : "string",
                    "description" : "Optional response format such as url or b64_json."
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/files" : {
      "get" : {
        "summary" : "List gateway-local Files",
        "operationId" : "get_v1_files",
        "tags" : [ "gateway" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "description" : "Lists gateway-local file objects owned by the current Distributed Key for conversation context, RAG and upload support.",
        "parameters" : [ {
          "name" : "purpose",
          "in" : "query",
          "required" : false,
          "description" : "Optional gateway-local purpose filter.",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "description" : "Number of Files to return, from 1 to 100. Defaults to 20.",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "after",
          "in" : "query",
          "required" : false,
          "description" : "Cursor id of the last File from the previous page.",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "order",
          "in" : "query",
          "required" : false,
          "description" : "Sort by creation time: asc or desc. Defaults to desc.",
          "schema" : {
            "type" : "string"
          }
        } ]
      },
      "post" : {
        "summary" : "Create a gateway-local File",
        "operationId" : "post_v1_files",
        "tags" : [ "gateway" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "description" : "Creates a gateway-local file object from multipart upload. Files support local content read and Vector Store attachment; this does not imply every upstream Files lifecycle feature.",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "required" : [ "file" ],
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary",
                    "description" : "File content to store in gateway-local file storage."
                  },
                  "purpose" : {
                    "type" : "string",
                    "description" : "Optional purpose such as assistants or batch; stored as metadata only for the functional service scope."
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/files/{fileId}/content" : {
      "get" : {
        "summary" : "Retrieve gateway-local File content",
        "operationId" : "get_v1_files_fileId_content",
        "tags" : [ "gateway" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "description" : "Returns raw gateway-local file content for downstream RAG or inspection flows.",
        "parameters" : [ {
          "name" : "fileId",
          "in" : "path",
          "required" : true,
          "description" : "Gateway-local File id.",
          "schema" : {
            "type" : "string"
          }
        } ]
      }
    },
...
  },
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "Distributed Key"
      }
    }
  }
}